Updated Geocoding API to support HTTPS
